Added a new method for rating images
Added a new method for converting image colors to YUV color space. The
colors are read from the website and are saved as Color in a list held
by each image. For setting up the priority the distance in YUV color
space is taken. The closer two colors are in YUV space the more similiar
they are. Currently all color with a distance smaller than 0.25 are
similar. That far to much; the range of values is between 0 and 1.